Add Mutex = std::mutex parameter to synchronized_value template
There are 6 types of mutexes in C++ (or rather 3, each in 2 variants):
- regular std::mutex   (plus std::timed_mutex)
- std::recursive_mutex (plus std::recursive_timed_mutex)
- std::shared_mutex    (plus std::shared_timed_mutex)

Let's the timed ones out scope for the time being. There is no support
for timed-limited locking of multiple mutexes at oncewith std::lock or
std::scoped_lock. Also what would it mean to `timed_loc(t, m1, m2)`?
Is `t` a total limit or per-lock-attempt limit? Should it be updated
with the passed wall-clock delta if `try_lock()` of a 2nd or next mutex
fails and next round of attempts is initiated? Too much trouble :)

The recursive mutex... is an abomination and an insult to proper code
architecture. If you don't control your threading model, sure, go ahead
with std::recursive_mutex and just kick that can down the road. It will
backfire on you or your colleague one day, be sure of that. This one
shouldn't exist at all, you can quote me.

The shared one however, this is a useful tool in producer-consumer
scenarios, it is very much legit and should be supported. Other
Lockables should not be excluded either. Since the dawn of C++
user-defined types have been first-class citizens, this shouldn't
change with the introduction of synchronized_value.