Add support for shared-locking vs. exclusive locking
`std::shared_mutex` can be locked in shared or exclusive mode. By
default it will it sill be locked exclusively, just like its own .lock()
method would do. However one might reasonably want to lock it in shared
mode too, using `.(un_|try_)lock_shared()` methods.

In order to allow it synchronized_value is augmented with with an new
`.share()` method. It will only be available if the Mutex class meets
the requirements of SharedLockable. It will return a wrapper which
`apply()` would accept as its parameter alongside regular
synchronized_value.

The shared locking always requires the lambda to accept const-reference
as parameter, as none of the possibly multiple readers is allowed to
modify the shared resource. Modification requires an exclusive lock,
obviously.