Make apply() compatible with using-declaration
4 apply() overloads are introduced with explicit synchronized_value and
shared_synchronized_value as first SV parameter, each in const and
non-const flavor.

This is necessary to avoid ambiguity with std::apply() from <tuple>
which accepts anything as its second parameter (i.e. the tuple) and thus
wins in overload resolution with the templated SynchronisedValueLike
that used to be the type of the second and any following parameters here.

This change not only enables `using BM::apply;` to be sprinkled around
in client code for sheer convenience, but also paves the way towards
moving the extended synchronized_value with shared locking support
directly into std namespace one day.

Pristine P0290r4 is free from this challenge because it doesn't have
shared_synchronized_value type and all parameters are explicitly and
concretely synchronized_value<T>&.

It is anticipated that std::apply from <tuple> would be augmented with
additional constraints on its tuple parameter in order to avoid this
ambiguity should such enriched synchronized_value with shared locking
support ever enter the standard.